Refactor move images and remove useless images
