+++ src/pages/super-admin/SuperAdminManage.tsx
@@ -2,9 +2,9 @@ import ImageRouteButton from '@components/common/button/ImageRouteButton';
 import AppContainer from '@components/common/container/AppContainer';
 import styled from '@emotion/styled';
 import useCustomNavigate from '@hooks/useCustomNavigate';
-import workspaceImage from '@resources/image/workspaceImage.png';
-import userImage from '@resources/image/userImage.png';
-import emailImage from '@resources/image/emailImage.png';
+import workspaceImage from '@resources/image/super-admin/workspaceImage.png';
+import userImage from '@resources/image/super-admin/userImage.png';
+import emailImage from '@resources/image/super-admin/emailImage.png';
 import { colFlex, rowFlex } from '@styles/flexStyles';
 
 const Container = styled.div`
